Removed parameters from nav constructor
Since apparently, user controls cannot have parameterized constructors as the designer is incapable of handling them when drag and dropped, I removed the parameter corresponding to the current Form that is passed in to the nav constructor and then used to close the current form.  I changed this to a method called "SetCurrentForm" and called it in the form's constructor.  It basically does the same exact thing.